requirements.txt
livekit>=1.0.3
livekit-agents==1.0.11
livekit-plugins-openai==1.0.11
livekit-plugins-deepgram==1.0.11
livekit-plugins-silero==1.0.11
python-dotenv~=1.0
pinecone>=3.0.0
openai
flask